requirements.txt
psycopg2-binary>=2.9
python-dotenv>=1.0.0
jinja2>=3.1
pytest>=7.0
